Reproductive Events Calendar

Create all variables related to the Reproductive Events calendar (see details). If this calendar was not included in the sample (or if some of the required information is not available), all variables will still be created, but all values will be NA.

vcal_reprod(dat, vcal_reprod_recodes = NULL)

Arguments

dat
A data file created by reshape_vcal_input_data()

vcal_reprod_recodes
Optional Character: the full path to a file called vcal_reprod_recodes.csv (contains all common and sample-specific recodes for the Reproductive Events calendar).

Details

The following variables will be created using case logic provided to the function dplyr::case_when(). Please note that case_when() returns NA through implicit logic: if a "case" exists and is not explicitly handled here, the value NA will be returned!

reprod_event Numeric: a recoded version of vcal_reprod (usually vcal_1 in the IR file). All values are a number, and sample-specific codes are harmonized using vcal_reprod_recodes.csv

caseid_cmc concatenates caseid and cmc_month (separated by _)

seq month index, where the earliest month gets 1 and the final month usually gets 80 (length of the original calendar string)

birth Logical: TRUE if vcal_reprod is recoded 100, FALSE if it is any other non-missing value.

preg Logical: TRUE if vcal_reprod is recoded 200, FALSE if it is any other non-missing value.

term Logical: TRUE if vcal_reprod is recoded 300, FALSE if it is any other non-missing value.

contr Logical: TRUE if vcal_reprod is recoded to a value between 1 and 90, FALSE if it is any other non-missing value.

eventpbt Logical: TRUE if birth, preg, or term is TRUE, FALSE if they are all FALSE and non-missing.

eventwfp Logical: TRUE if eventpbt or contr is TRUE, FALSE if both are FALSE and non-missing.

switch Logical: TRUE if vcal_reprod is recoded 1:90 and the previous month's value for vcal_reprod is recoded 0:90 but not the same value as the current month. FALSE in all other cases where both the current and previous month are available (e.g. excludes the first month and any missing value) and either falls beyond the required range.

switch_new Logical: TRUE if switch is TRUE and vcal_reprod in the previous month not recoded 0. FALSE if switch is TRUE and vcal_reprod in the previous month is recoded 0.

disc_event see vcal_discont()

total_preg Integer: the total number of months per person where preg is TRUE (NA if preg is not available). All months for the person reflect the same total (this is not a cumulative sum).

count_birth Integer: the total number of months per person where birth is TRUE (NA if birth is not available). All months for the person reflect the same total (this is not a cumulative sum).

count_term Integer: the total number of months per person where term is TRUE (NA if term is not available). All months for the person reflect the same total (this is not a cumulative sum).

contr_duration Integer: the total number of months per person where contr is TRUE (NA if contr is not available). All months for the person reflect the same total (this is not a cumulative sum).

trunc Logical: TRUE for all months in a continuous pregnancy if the outcome of the pregnancy (birth or termination) happened after the date of the interview. FALSE for all months in a continuous pregnancy otherwise. Not available (NA) for all months that were not part of a continuous pregnancy.

prfirst Logical: TRUE for all months in a continuous pregnancy that overlaps with the first month in the recall timeline. FALSE for all months in a continuous pregnancy otherwise. Not available (NA) for months that were not part of a continuous pregnancy.

preg_length Integer: the total number of months included in a single continuous pregnancy where both trunc and prfirst are FALSE. All months in the pregnancy reflect the same total (this is not a cumulative sum). Not available (NA) for months that were not part of a continuous pregnancy.

preg_flag Logical: TRUE if preg_length is longer than 9 months. Not available (NA) for months where preg_length is not available.

Author

Matt Gunther