Pregnancy Termination Calendar

Create all variables related to the Pregnancy Termination calendar (see details). If this calendar was not included in the sample (or if some of the required information is not available), all variables will still be created, but all values will be NA.

vcal_term(dat)

Arguments

dat

A data file created by vcal_reprod() (may be passed to any other function starting with "vcal" first).

Details

The following variables will be created using case logic provided to the function dplyr::case_when(). Please note that case_when() returns NA through implicit logic: if a "case" exists and is not explicitly handled here, the value NA will be returned!

abort Logical: TRUE if vcal_term is "A" and vcal_reprod indicates that the woman experienced pregnancy termination (value 300). FALSE if vcal_term is any other non-missing character (months of termination only). NA if vcal_reprod shows no termination, if vcal_term is not included in the sample, or if the response in vcal_term is missing.

abort_total Integer: total number of months per person where abort is TRUE. Zero if none of the woman's terminations were abortions. NA if the woman never experienced pregnancy termination, or if vcal_term was not included in the sample. All months for a given woman reflect the same total (this is not a cumulative sum).

miscar Logical: TRUE if vcal_term is "M" or "C" and vcal_reprod indicates that the woman experienced pregnancy termination (value 300). FALSE if vcal_term is any other non-missing character (months of termination only). NA if vcal_reprod shows no termination, if vcal_term is not included in the sample, or if the response in vcal_term is missing.

miscar_total Integer: total number of months per person where miscar is TRUE. Zero if none of the woman's terminations were miscarriages NA if the woman never experienced pregnancy termination, or if vcal_term was not included in the sample. All months for a given woman reflect the same total (this is not a cumulative sum).

sbirth Logical: TRUE if vcal_term is "S" and vcal_reprod indicates that the woman experienced pregnancy termination (value 300). FALSE if vcal_term is any other non-missing character (months of termination only). NA if vcal_reprod shows no termination, if vcal_term is not included in the sample, or if the response in vcal_term is missing.

sbirth_total Integer: total number of months per person where sbirth is TRUE. Zero if none of the woman's terminations were stillbirths. NA if the woman never experienced pregnancy termination, or if vcal_term was not included in the sample. All months for a given woman reflect the same total (this is not a cumulative sum).

Author

Matt Gunther