Reproductive Events Calendar

Create all variables related to the Reproductive Events calendar (see details). If this calendar was not included in the sample (or if some of the required information is not available), all variables will still be created, but all values will be NA. See note regarding NA values for contraceptive use.

vcal_reprod(dat, vcal_reprod_recodes = NULL)

Arguments

dat

A data file created by reshape_vcal_input_data()

vcal_reprod_recodes

Optional Character: the full path to a file called vcal_reprod_recodes.csv (contains all common and sample-specific recodes for the Reproductive Events calendar).

Details

The following variables will be created using case logic provided to the function dplyr::case_when(). Please note that case_when() returns NA through implicit logic: if a "case" exists and is not explicitly handled here, the value NA will be returned!

vcal_reprod Numeric: a recoded version of the Reproductive Events calendar (usually vcal_1 in the IR file). All values are a number; both common and sample-specific codes are harmonized using vcal_reprod_recodes.csv

vcal_reprod_dhs Character: The alphanumeric DHS codes used in the Reproductive Events calendar (usually vcal_1 in the IR file). These are not likely to be published by IPUMS DHS, but they may be useful for quality checking.

caseid_cmc concatenates caseid and cmc_month (separated by _)

seq month index, where the earliest month gets 1 and the final month usually gets 80 (length of the original calendar string)

birth Logical: TRUE if vcal_reprod is 100, FALSE if it is any other non-missing value (NA otherwise).

preg Logical: TRUE if vcal_reprod is 200, FALSE if it is any other non-missing value (NA otherwise).

term Logical: TRUE if vcal_reprod is 300, FALSE if it is any other non-missing value (NA otherwise).

contr Logical: TRUE if vcal_reprod is a value between 1 and 90, FALSE if it is any other non-missing value (NA otherwise).

eventpbt Logical: TRUE if birth, preg, or term is TRUE, FALSE if they are all FALSE and non-missing (NA otherwise).

eventwfp Logical: TRUE if eventpbt or contr is TRUE, FALSE if both are FALSE and non-missing (NA otherwise).

contr_start Logical: TRUE if vcal_reprod is between 1:90 and the previous month's value for vcal_reprod is between 0:90 and different from the current month. FALSE if vcal_reprod is between 1:90 and the previous month's value for vcal_reprod is the same as the current month. NA in any other case (e.g. the first month of the recall timeline is NA; months where vcal_reprod is beyond 1:90 are NA; months where the previous month vcal_reprod is beyond 0:90 are NA).

contr_change Logical: TRUE if contr_start is TRUE and vcal_reprod in the previous month is not 0 (i.e. the method was started following a month of non-use of any method). FALSE if contr_start is TRUE and vcal_reprod in the previous month is 0. NA if contr_start is NA.

contr_stop see vcal_discont()

preg_total Integer: the total number of months per person where preg is TRUE (NA if all values for preg are NA). All months for the person reflect the same total (this is not a cumulative sum).

birth_total Integer: the total number of months per person where birth is TRUE (NA if all values for birth are NA). All months for the person reflect the same total (this is not a cumulative sum).

term_total Integer: the total number of months per person where term is TRUE (NA if all values for term are NA). All months for the person reflect the same total (this is not a cumulative sum).

contr_total Integer: the total number of months per person where contr is TRUE (NA if all values for contr are NA). All months for the person reflect the same total (this is not a cumulative sum).

preg_rc Logical: TRUE for all months in a continuous pregnancy if the outcome of the pregnancy (birth or termination) happened after the date of the interview. FALSE for all months in a continuous pregnancy otherwise. NA for all months that were not part of a continuous pregnancy.

preg_lc Logical: TRUE for all months in a continuous pregnancy (including months of birth or termination) that overlaps with the first month in the recall timeline. FALSE for all months in a continuous pregnancy otherwise. NA for months that were not part of a continuous pregnancy.

preg_length Integer: the total number of months included in a single continuous pregnancy (including months of birth or termination) where both preg_rc and preg_lc are FALSE. All months in the pregnancy reflect the same total (this is not a cumulative sum). NA for months that were not part of a continuous pregnancy.

preg_long Logical: TRUE if preg_length is longer than 9 months. NA for months where preg_length is not available.

Note

All samples seem to include a Reproductive Events calendar, but some samples do no include responses for monthly contraceptive use (you will find only "pregnancy", "birth", or "termination" events listed). In these samples, it is common to see blank values in the original DHS string; we interpret these blanks as months in which a reproductive event did not occur and recode them as the value 0. For example, if the Reproductive Events calendar for a woman in one of these samples contains only blanks, we assume that she has zero "pregnant" months. These samples include no NA values for missing data because missing data are indistinguishable from the original blank DHS values (possibly except for blanks in the middle of a continuous pregnancy).

The majority of samples do contain responses for monthly contraceptive use, and these samples do contain NA values for missing data (e.g. a woman could not recall what method she used in a particular month). For example, if the Reproductive Events calendar for a woman in one of these samples contains only missing values, we assume that the woman was NIU (not in universe) for the Reproductive Events calendar section of the survey (or else declined to provide any responses); all months for such a woman will be marked NA (e.g. the calculated sum for her total number of "pregnant" months will be NA).

Because NA values may represent both "NIU" and "missing" depending on the sample, the recommended value label for NA is for monthly indicators is "No response or NIU". The interpretation for personal totals or event summaries (e.g. length of pregnancy) can be treated as "NIU" where the universe includes relevant months (e.g. months that were part of a single, continuous pregnancy).

Author

Matt Gunther